高抬腿主页点击后的其他逻辑都实现

+++ app/src/main/java/com/fjp/skeletalmuscle/data/model/bean/result/LiftLegTrendResult.kt
@@ -5,6 +5,6 @@ package com.fjp.skeletalmuscle.data.model.bean.result
  *Time:2024/12/1
  *Description:
  */
-data class LiftLegTrendResult(val avg_left_degree: Int, val avg_right_degree: Int, val trend: List<LeftTrend>)
+data class LiftLegTrendResult(val avg_left_degree: Double, val avg_right_degree: Double, val trend: List<LeftTrend>)
 
 data class LeftTrend(val time: String, val left_degree: Double, val right_degree: Double)
